commit 19-09 | 20-08

+++ adm/userManagment.php
@@ -1,34 +1,40 @@
 <?php
+    session_start();
     include 'inc/header.inc.php';
     include 'class/user.class.php';
     
     $user = new User();
+    
 ?>
 
-<div class="bg-info d-grid">
+
+<div class="bgMiddle sizeIndex justify-content-start">
 
     <h1>Gerenciamento de Usuários</h1>
 
-    <hr>
+    <div class="d-flex justify-content-center clearfix">
+        <div class="d-flex justify-content-center align-center text-center mb-3 me-5">
+            <button class="btn btn-success"><a href="addUser.php">ADICIONAR</a></button>
+        </div>
 
-    <div class="d-flex justify-content-center align-center text-center mb-3">
-        <button class="btn btn-success"><a href="addUser.php">ADICIONAR</a></button>
+        <div class="d-flex justify-content-center align-center text-center mb-3 ms-5">
+            <button class="btn btn-success"><a href="index.php">Voltar</a></button>
+        </div>
     </div>
 
+    <div class="d-flex justify-content-center align-center text-center size">
 
-    <div = class="d-flex justify-content-center align-center text-center">
-
-        <table class="table-bordered mb-5" width="99%">
+        <table class="table table-bordered" width="99%">
             <tr>
-                <th class="bg-light">ID</th>
-                <th class="bg-secondary">Nome</th>
-                <th class="bg-light">Data Nascimento</th>
-                <th class="bg-secondary">Nick</th>
-                <th class="bg-light">Senha</th>
-                <th class="bg-secondary">Email</th>
-                <th class="bg-light">Telefone</th>
-                <th class="bg-secondary">Permissão</th>
-                <th class="bg-light"></th>
+                <th class="fw-bold">ID</th>
+                <th class="fw-bold">Nome</th>
+                <th class="fw-bold">Data Nascimento</th>
+                <th class="fw-bold">Nick</th>
+                <!--<th class="fw-bold">Senha</th>-->
+                <th class="fw-bold">Email</th>
+                <th class="fw-bold">Telefone</th>
+                <th class="fw-bold">Permissão</th>
+                <th class="fw-bold"></th>
             </tr>
 
             <?php
@@ -38,17 +44,17 @@
 
             <tbody class="">
                     <tr>
-                        <td class="p-2 bg-light"><?php echo $item['id_user']; ?></td>
-                        <td class="p-2 bg-secondary"><?php echo $item['nome']; ?></td>
-                        <td class="p-2 bg-light"><?php echo implode("/", array_reverse(explode("-", $item['dtNasc']))); ?></td>
-                        <td class="p-2 bg-secondary"><?php echo $item['nick']; ?></td>
-                        <td class="p-2 bg-light"><?php echo $item['senha']; ?></td>
-                        <td class="p-2 bg-secondary"><?php echo $item['email'] ;?></td>
-                        <td class="p-2 bg-light"><?php echo $item['telefone']; ?></td>
-                        <td class="p-2 bg-secondary"><?php echo $item['permissao']; ?></td>
-                        <td class="p-2 bg-light">
+                        <td class="p-2"><?php echo $item['id_user']; ?></td>
+                        <td class="p-2"><?php echo $item['nome']; ?></td>
+                        <!--<td class="p-2"><?php //echo implode("/", array_reverse(explode("-", $item['dtNasc']))); ?></td>-->
+                        <td class="p-2"><?php echo $item['nick']; ?></td>
+                        <td class="p-2"><?php echo $item['senha']; ?></td>
+                        <td class="p-2"><?php echo $item['email'] ;?></td>
+                        <td class="p-2"><?php echo $item['telefone']; ?></td>
+                        <td class="p-2"><?php echo $item['permissao']; ?></td>
+                        <td class="p-2">
                             <button class="btn btn-dark btn-sm"><a href="editUser.php?id_user=<?php echo $item['id_user']?>">EDITAR</a></button>
-                            <button class="btn btn-dark btn-sm"><a href="deleteUser.php?id_user=<?php echo $item['id_user']?>">EXCLUIR</a></button>
+                            <button class="btn btn-dark btn-sm"><a href="deleteUser.php?id_user=<?php echo $item['id_user']?>" onclick="return confirm('Você tem CERTEZA disso?')">EXCLUIR</a></button>
                         </td>
                     </tr>
                 </tbody>
@@ -59,7 +65,6 @@
         </table>
 
     </div>
-
 </div>
